Revert "Bratseth/instances in deployment xml rebased"
