+++ controller-server/src/test/java/com/yahoo/vespa/hosted/controller/ControllerTest.java
@@ -72,6 +72,7 @@ public class ControllerTest {
     @Test
     public void testDeployment() {
         // Setup system
+        ApplicationController applications = tester.controller().applications();
         ApplicationPackage applicationPackage = new ApplicationPackageBuilder()
                 .environment(Environment.prod)
                 .region("us-west-1")
@@ -752,7 +753,7 @@ public class ControllerTest {
             tester.deployCompletely(application, applicationPackage);
             fail("Expected exception");
         } catch (IllegalArgumentException e) {
-            assertEquals("Endpoint 'default' in instance 'default' cannot contain regions in different clouds: [aws-us-east-1, us-west-1]", e.getMessage());
+            assertEquals("Endpoint 'default' cannot contain regions in different clouds: [aws-us-east-1, us-west-1]", e.getMessage());
         }
 
         var applicationPackage2 = new ApplicationPackageBuilder()
@@ -765,7 +766,7 @@ public class ControllerTest {
             tester.deployCompletely(application, applicationPackage2);
             fail("Expected exception");
         } catch (IllegalArgumentException e) {
-            assertEquals("Endpoint 'foo' in instance 'default' cannot contain regions in different clouds: [aws-us-east-1, us-west-1]", e.getMessage());
+            assertEquals("Endpoint 'foo' cannot contain regions in different clouds: [aws-us-east-1, us-west-1]", e.getMessage());
         }
     }
