+++ controller-server/src/test/java/com/yahoo/vespa/hosted/controller/restapi/application/ApplicationApiTest.java
@@ -113,18 +113,7 @@ public class ApplicationApiTest extends ControllerContainerTest {
                                                "-----END PUBLIC KEY-----\n";
     private static final String quotedPemPublicKey = pemPublicKey.replaceAll("\\n", "\\\\n");
 
-    private static final ApplicationPackage applicationPackageDefault = new ApplicationPackageBuilder()
-            .instances("default")
-            .environment(Environment.prod)
-            .globalServiceId("foo")
-            .region("us-central-1")
-            .region("us-east-3")
-            .region("us-west-1")
-            .blockChange(false, true, "mon-fri", "0-8", "UTC")
-            .build();
-
-    private static final ApplicationPackage applicationPackageInstance1 = new ApplicationPackageBuilder()
-            .instances("instance1")
+    private static final ApplicationPackage applicationPackage = new ApplicationPackageBuilder()
             .environment(Environment.prod)
             .globalServiceId("foo")
             .region("us-central-1")
@@ -232,7 +221,7 @@ public class ApplicationApiTest extends ControllerContainerTest {
         addUserToHostedOperatorRole(HostedAthenzIdentities.from(HOSTED_VESPA_OPERATOR));
 
         // POST (deploy) an application to a zone - manual user deployment (includes a content hash for verification)
-        MultiPartStreamer entity = createApplicationDeployData(applicationPackageInstance1, true);
+        MultiPartStreamer entity = createApplicationDeployData(applicationPackage, true);
         tester.assertResponse(request("/application/v4/tenant/tenant1/application/application1/environment/dev/region/us-west-1/instance/instance1/deploy", POST)
                                       .data(entity)
                                       .header("X-Content-Hash", Base64.getEncoder().encodeToString(Signatures.sha256Digest(entity::data)))
@@ -252,7 +241,7 @@ public class ApplicationApiTest extends ControllerContainerTest {
         controllerTester.jobCompletion(JobType.component)
                         .application(id)
                         .projectId(screwdriverProjectId)
-                        .uploadArtifact(applicationPackageInstance1)
+                        .uploadArtifact(applicationPackage)
                         .submit();
 
         // ... systemtest
@@ -316,7 +305,6 @@ public class ApplicationApiTest extends ControllerContainerTest {
 
         // POST (create) another application
         ApplicationPackage applicationPackage = new ApplicationPackageBuilder()
-                .instances("instance1")
                 .environment(Environment.prod)
                 .region("us-west-1")
                 .build();
@@ -593,7 +581,6 @@ public class ApplicationApiTest extends ControllerContainerTest {
 
         // Second attempt has a service under a different domain than the tenant of the application, and fails.
         ApplicationPackage packageWithServiceForWrongDomain = new ApplicationPackageBuilder()
-                .instances("instance1")
                 .environment(Environment.prod)
                 .athenzIdentity(com.yahoo.config.provision.AthenzDomain.from(ATHENZ_TENANT_DOMAIN_2.getName()), AthenzService.from("service"))
                 .region("us-west-1")
@@ -606,7 +593,6 @@ public class ApplicationApiTest extends ControllerContainerTest {
 
         // Third attempt finally has a service under the domain of the tenant, and succeeds.
         ApplicationPackage packageWithService = new ApplicationPackageBuilder()
-                .instances("instance1")
                 .environment(Environment.prod)
                 .athenzIdentity(com.yahoo.config.provision.AthenzDomain.from(ATHENZ_TENANT_DOMAIN.getName()), AthenzService.from("service"))
                 .region("us-west-1")
@@ -720,7 +706,6 @@ public class ApplicationApiTest extends ControllerContainerTest {
         tester.computeVersionStatus();
         createAthenzDomainWithAdmin(ATHENZ_TENANT_DOMAIN, USER_ID);
         ApplicationPackage applicationPackage = new ApplicationPackageBuilder()
-                .instances("instance1")
                 .globalServiceId("foo")
                 .region("us-west-1")
                 .region("us-east-3")
@@ -729,7 +714,7 @@ public class ApplicationApiTest extends ControllerContainerTest {
         // Create tenant and deploy
         ApplicationId id = createTenantAndApplication();
         long projectId = 1;
-        MultiPartStreamer deployData = createApplicationDeployData(Optional.of(applicationPackage), false);
+        MultiPartStreamer deployData = createApplicationDeployData(Optional.empty(), false);
         startAndTestChange(controllerTester, id, projectId, applicationPackage, deployData, 100);
 
         // us-west-1
@@ -792,7 +777,6 @@ public class ApplicationApiTest extends ControllerContainerTest {
         tester.computeVersionStatus();
         createAthenzDomainWithAdmin(ATHENZ_TENANT_DOMAIN, USER_ID);
         ApplicationPackage applicationPackage = new ApplicationPackageBuilder()
-                .instances("instance1")
                 .region("us-west-1")
                 .region("us-east-3")
                 .region("eu-west-1")
@@ -869,7 +853,7 @@ public class ApplicationApiTest extends ControllerContainerTest {
                                        new com.yahoo.vespa.hosted.controller.api.identifiers.ApplicationId("application1"));
 
         // POST (deploy) an application to a prod zone - allowed when project ID is not specified
-        MultiPartStreamer entity = createApplicationDeployData(applicationPackageInstance1, true);
+        MultiPartStreamer entity = createApplicationDeployData(applicationPackage, true);
         tester.assertResponse(request("/application/v4/tenant/tenant1/application/application1/environment/prod/region/us-central-1/instance/instance1/deploy", POST)
                                       .data(entity)
                                       .screwdriverIdentity(SCREWDRIVER_ID),
@@ -901,7 +885,6 @@ public class ApplicationApiTest extends ControllerContainerTest {
 
         // Deploy
         ApplicationPackage applicationPackage = new ApplicationPackageBuilder()
-                .instances("instance1")
                 .region("us-east-3")
                 .build();
         ApplicationId id = createTenantAndApplication();
@@ -921,7 +904,6 @@ public class ApplicationApiTest extends ControllerContainerTest {
 
         // New zone is added before us-east-3
         applicationPackage = new ApplicationPackageBuilder()
-                .instances("instance1")
                 .globalServiceId("foo")
                 // These decides the ordering of deploymentJobs and instances in the response
                 .region("us-west-1")
@@ -1074,7 +1056,7 @@ public class ApplicationApiTest extends ControllerContainerTest {
         configServer.throwOnNextPrepare(new ConfigServerException(new URI("server-url"), "Failed to prepare application", ConfigServerException.ErrorCode.INVALID_APPLICATION_PACKAGE, null));
         
         // POST (deploy) an application with an invalid application package
-        MultiPartStreamer entity = createApplicationDeployData(applicationPackageInstance1, true);
+        MultiPartStreamer entity = createApplicationDeployData(applicationPackage, true);
         tester.assertResponse(request("/application/v4/tenant/tenant1/application/application1/environment/dev/region/us-west-1/instance/instance1/deploy", POST)
                                       .data(entity)
                                       .userIdentity(USER_ID),
@@ -1194,7 +1176,7 @@ public class ApplicationApiTest extends ControllerContainerTest {
                               200);
 
         // Deploy to an authorized zone by a user tenant is disallowed
-        MultiPartStreamer entity = createApplicationDeployData(applicationPackageDefault, true);
+        MultiPartStreamer entity = createApplicationDeployData(applicationPackage, true);
         tester.assertResponse(request("/application/v4/tenant/tenant1/application/application1/environment/prod/region/us-west-1/instance/default/deploy", POST)
                                       .data(entity)
                                       .userIdentity(USER_ID),
@@ -1607,7 +1589,7 @@ public class ApplicationApiTest extends ControllerContainerTest {
     }
 
     private MultiPartStreamer createApplicationDeployData(Optional<ApplicationPackage> applicationPackage,
-                                                          Optional<ApplicationVersion> applicationVersion, boolean deployDirectly) {
+                                                   Optional<ApplicationVersion> applicationVersion, boolean deployDirectly) {
         MultiPartStreamer streamer = new MultiPartStreamer();
         streamer.addJson("deployOptions", deployOptions(deployDirectly, applicationVersion));
         applicationPackage.ifPresent(ap -> streamer.addBytes("applicationZip", ap.zippedContent()));
